+++ pyabc/population.py
@@ -1,3 +1,13 @@
+"""
+Particles and Populations
+=========================
+
+A particle contains the sampled parameters and simulated data.
+A population gathers all particles collected in one SMC
+iteration.
+"""
+
+
 from typing import List, Callable
 import pandas as pd
 from pyabc.parameters import Parameter
@@ -15,30 +25,31 @@ class Particle:
     ----------
 
     m: int
-        The model nr
+        The model index.
 
     parameter: Parameter
-        The model specific parameter
+        The model specific parameter.
 
-    weight: float, 0 < weight < 1
-        The weight of the particle
+    weight: float, 0 <= weight <= 1
+        The weight of the particle.
 
-    distances: List[float]
+    accepted_sum_stats: List[dict]
+        List of accepted summary statistics.
+        This list is usually of length 1. This list is longer only if more
+        than one sample is taken for a particle.
+        This list has length 0 if the particle is rejected.
+
+    accepted_distances: List[float]
         A particle can contain more than one sample.
         If so, the distances of the individual samples
         are stored in this list. In the most common case of a single
         sample, this list has length 1.
 
-    accepted_sum_stats
-        List of accepted summary statistics which describe the particle
-        This list is usually of length 1. This list is longer only if more
-        than one sample is taken for a particle.
-        This list has length 0 if the particle is rejected.
+    rejected_sum_stats: List[dict]
+        List of rejected summary statistics.
 
-    all_sum_stats: List[dict]
-        List of all summary statistics generated during the creation of this
-        particle (also when they led to rejection).
-        This list is non-empty also for rejected particles.
+    rejected_distances: List[float]
+        List of rejected distances.
 
     accepted: bool
         True if particle was accepted, False if not.
@@ -56,43 +67,28 @@ class Particle:
         multiplying the weights with the model probabilities.
     """
 
-    def __init__(self, m: int,
+    def __init__(self,
+                 m: int,
                  parameter: Parameter,
                  weight: float,
-                 accepted_distances: List[float],
                  accepted_sum_stats: List[dict],
-                 all_sum_stats: List[dict] = None,
+                 accepted_distances: List[float],
+                 rejected_sum_stats: List[dict] = None,
+                 rejected_distances: List[float] = None,
                  accepted: bool = True):
 
         self.m = m
         self.parameter = parameter
         self.weight = weight
-        self.accepted_distances = accepted_distances
         self.accepted_sum_stats = accepted_sum_stats
-        self.all_sum_stats = all_sum_stats
+        self.accepted_distances = accepted_distances
+        if rejected_sum_stats is None:
+            rejected_sum_stats = []
+        self.rejected_sum_stats = rejected_sum_stats
+        if rejected_distances is None:
+            rejected_distances = []
+        self.rejected_distances = rejected_distances
         self.accepted = accepted
-
-    def __getitem__(self, item):
-        return getattr(self, item)
-
-    def __setitem__(self, key, value):
-        setattr(self, key, value)
-
-    def __eq__(self, other):
-        for key in ["m", "parameter", "weight", "accepted_distances",
-                    "accepted_sum_stats", "all_sum_stats"]:
-            if self[key] != other[key]:
-                return False
-        return True
-
-    def copy(self):
-        return self.__class__(self.m,
-                              self.parameter,
-                              self.weight,
-                              self.accepted_distances,
-                              self.accepted_sum_stats,
-                              self.all_sum_stats,
-                              self.accepted)
 
 
 class Population:
@@ -103,7 +99,7 @@ class Population:
     """
 
     def __init__(self, particles: List[Particle]):
-        self._list = [particle.copy() for particle in particles]
+        self._list = particles.copy()
         self._model_probabilities = None
         self._normalize_weights()